requirements.txt
asgiref==3.2.3
attrs==19.3.0
azure-common==1.1.24
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
Babel==2.7.0
backcall==0.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
dateutils==0.6.7
decorator==4.4.1
defusedxml==0.6.0
Django==3.0.2
django-allauth==0.41.0
django-credit-cards==0.4.1
django-filter==2.2.0
django-hstore==1.4.2
django-phonenumber-field==3.0.1
django-rest-auth==0.9.5
django-rest-framework-braces==0.3.4
django-sendgrid-v5==0.8.1
django-storages==1.9.1
django-webpack-loader==0.6.0
djangorestframework==3.11.0
djangorestframework-hstore==1.3
future==0.18.2
idna==2.8
importlib-metadata==1.4.0
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
Markdown==3.1.1
more-itertools==8.1.0
oauthlib==3.1.0
packaging==20.0
parso==0.5.2
pexpect==4.7.0
phonenumbers==8.11.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
prompt-toolkit==3.0.2
psycopg2==2.8.4
psycopg2-binary==2.8.4
ptyprocess==0.6.0
py==1.8.1
pycparser==2.19
Pygments==2.5.2
pyparsing==2.4.6
pytest==5.3.2
pytest-django==3.8.0
python-dateutil==2.8.1
python-devtools==2
python-dotenv==0.10.3
python-http-client==3.2.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
sendgrid==6.1.0
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.8
zipp==1.0.0
